Folder structure cleanup
Sources nicely in their respective folders mirroring the solution file.

+++ bus/rom/rom.cpp
@@ -12,9 +12,9 @@ rom::rom() {
 rom::~rom() {
 }
 
-void rom::set_rom_data(byte *data, int size) {
+void rom::set_rom_data(byte *data, size_t size) {
 	this->romdata = data;
-	this->romsize = size;
+	this->romsize = (int)size;
 	// extra logic for 16kB / 32kB roms.
 	if (size == 16384) devicemask = 0xBFFF; else devicemask = 0xFFFF;
 }
